migrate useInitMachine, and rename to useInitAuthenticatorMachine
