+++ packages/react-core/src/Authenticator/hooks/index.ts
@@ -1,5 +1,6 @@
 export * from './useAuthenticator';
 export * from './useAuthenticatorRoute';
+export * from './useInitAuthenticatorMachine';
 export { resolveAuthenticatorComponents } from './utils';
 
 export {
