+++ packages/react-core/src/Authenticator/hooks/useInitAuthenticatorMachine/index.ts
@@ -1,0 +1,1 @@
+export { default as useInitAuthenticatorMachine } from './useInitAuthenticatorMachine';
