+++ packages/react/src/components/Authenticator/Authenticator.tsx
@@ -5,6 +5,7 @@ import {
   AuthenticatorProvider as Provider,
   useAuthenticator,
   UseAuthenticator,
+  useInitAuthenticatorMachine,
 } from '@aws-amplify/ui-react-core';
 import {
   CustomComponentsContext,
@@ -29,20 +30,6 @@ export type AuthenticatorProps = Partial<
     }
 >;
 
-// Utility hook that sends init event to the parent provider
-function useInitMachine(data: AuthenticatorMachineOptions) {
-  const { route, initializeMachine } = useAuthenticator(({ route }) => [route]);
-
-  const hasInitialized = React.useRef(false);
-  React.useEffect(() => {
-    if (!hasInitialized.current && route === 'setup') {
-      initializeMachine(data);
-
-      hasInitialized.current = true;
-    }
-  }, [initializeMachine, route, data]);
-}
-
 // `AuthenticatorInternal` exists to give access to the context returned via `useAuthenticator`,
 // which allows the `Authenticator` to just return `children` if a user is authenticated.
 // Once the `Provider` is removed from the `Authenticator` component and exported as
@@ -64,7 +51,7 @@ export function AuthenticatorInternal({
     ({ route, signOut, user }) => [route, signOut, user]
   );
 
-  useInitMachine({
+  useInitAuthenticatorMachine({
     initialState,
     loginMechanisms,
     services,
